Extracting effects outside components
The bssic case works.

The `useEffect()` docs explained that my effect
functions could return either a named function or an arrow function,
which meant that `this` binding probably doesn't affect effect
registration, and the code should work whether the effect gets defined
inside or outside the component.

My logging effect only needs APIs and data from outside the component.

If my effect captured some state from the component, this function
extraction wouldn't work. The external function wouldn't have access to
the captured variable any more, unless I write explicit `this` binding
(and that might not work in every case: I don't know how `this` is bound
for components by React generally, so I can't assume it binds as
expected. This differs from class-based components. In those I know
`this` gets bound to the instance for lifecycle methods, because I
have used it and remember from documentation. I don't know if the
same holds for functional components and hooks. This (ahem) seems
worth an experiement, but I'm not sure I have the will for it right
now.