Consume Flickr public feed API in useEffect
Adds an Effect which requests the public feed from Flickr. It uses the
`jsonp` library which provides an API on making JSONP calls. Adds a
callback to process the data from the API (add it to state) and passes
this to the `jsonp()` function.

When the state updates with the new data, render an ordered list
containing images and descriptions. The images link to the flickr.com
page.

This proves the typescript, react, jsonp, flickr stack can work.